Add CopiableHash module and adjust ZobristHash implementation
This commit introduces a new module, CopiableHash, and adjusts the ZobristHash implementation. Mainly, the ability to double add or empty remove elements in debug builds are now tested using a CopiableHash structure instead of a HashSet. The addition of the dev-dependencies and feature flag "check_set_behavior" in the Cargo.toml file is to facilitate this change. Also, a new example showing a chess board hash implementation has been added.